refactor: consolidate chat types and fix layer/hook violations
- Move chat types (ChatMessage, ChatErrorCode, ChatLoadingPhase, ChatSession, ChatActionResult) from domain/chat/types.ts to domain/types.ts for proper layer encapsulation
- Remove re-export from domain/chat/types.ts; delete the file after consolidation
- Update ChatPromptPayload.messages to use ChatMessage[] array
- Replace magic number 429 with HTTP_STATUS_TOO_MANY_REQUESTS constant in chatAction.ts
- Convert getRedis() to lazy singleton pattern in tokenStore.ts
- Export resetRedisForTest() for test isolation in tokenStore.ts
- Move useQueryClient/useQuery declarations after useRef hook in useChat.ts (fix hook order)
- Extract ChatSessionLoadResult interface in chatStorage.ts for clarity
