+++ src/domain/types.ts
@@ -642,15 +642,25 @@ export type DashboardTimeframe = '15Min' | '1Hour' | '1Day';
 
 // ─── Chat ─────────────────────────────────────────────────────────────────────
 
-export type {
-    ChatMessage,
-    ChatErrorCode,
-    ChatLoadingPhase,
-    ChatSession,
-    ChatActionResult,
-} from './chat/types';
+export interface ChatMessage {
+    role: 'user' | 'model';
+    content: string;
+}
+
+export type ChatLoadingPhase = 'analyzing' | 'generating';
+
+export interface ChatSession {
+    messages: ChatMessage[];
+    savedAt: number; // Unix timestamp (ms)
+}
+
+export type ChatErrorCode = 'token_exhausted' | 'rate_limited' | 'server_error';
+
+export type ChatActionResult =
+    | { ok: true; message: string; remainingTokens: number }
+    | { ok: false; error: ChatErrorCode };
 
 export interface ChatPromptPayload {
     systemPrompt: string;
-    messages: Array<{ role: 'user' | 'model'; content: string }>;
+    messages: ChatMessage[];
 }
